requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
databricks-sql-connector>=3.0.0